Make mbtileserver 'go get'-able
This commit refactors the templates and statics to the assets
http.FileSystem defined in package handler instead of relying on the
real filesystem. For the default build all needed assets are within the
executable.

When built with '-tags dev', the executable must still be run with the
project repo being the current working directory, otherwise the
templates creation will fail.